Add shared memory /dev/shm volume (#269)
By default the docker has limit in 64MB for `/dev/shm` set. Which later used for postgres for it's shared memory settings.  That means - whatever we set in the `user.conf` for shared_buffers - won't be used in full capacity and Postgres will with following errors 
```
ERROR:  could not resize shared memory segment "/PostgreSQL.1624162460" to 50438144 bytes: No space left on device
```

This PR should add default `EmptyFolder` in-memory (tmpfs) in place of `/dev/shm` to prevent this.

Keep in mind that this was not properly tested and need some additional love and care before can be merged anywhere